requirements.txt
autopep8==1.6.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Deprecated==1.2.13
Flask==2.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.2
jwt==1.3.1
MarkupSafe==2.1.1
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.20.0
python-jwt==3.3.2
SQLAlchemy==1.4.36
toml==0.10.2
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
